+++ src/cargo/core/compiler/mod.rs
@@ -93,8 +93,8 @@ use crate::core::{Feature, PackageId, Target, Verbosity};
 use crate::util::errors::{CargoResult, VerboseError};
 use crate::util::interning::InternedString;
 use crate::util::machine_message::{self, Message};
-use crate::util::toml::TomlDebugInfo;
-use crate::util::toml::TomlTrimPaths;
+use crate::util::toml::schema::TomlDebugInfo;
+use crate::util::toml::schema::TomlTrimPaths;
 use crate::util::{add_path_args, internal, iter_join_onto, profile};
 use cargo_util::{paths, ProcessBuilder, ProcessError};
 use rustfix::diagnostics::Applicability;
