+++ src/cargo/core/manifest.rs
@@ -18,7 +18,7 @@ use crate::core::{Dependency, PackageId, PackageIdSpec, SourceId, Summary};
 use crate::core::{Edition, Feature, Features, WorkspaceConfig};
 use crate::util::errors::*;
 use crate::util::interning::InternedString;
-use crate::util::toml::{TomlManifest, TomlProfiles};
+use crate::util::toml::schema::{TomlManifest, TomlProfiles};
 use crate::util::{short_hash, Config, Filesystem, RustVersion};
 
 pub enum EitherManifest {
