+++ src/cargo/core/mod.rs
@@ -14,7 +14,7 @@ pub use self::workspace::{
     find_workspace_root, resolve_relative_path, MaybePackage, Workspace, WorkspaceConfig,
     WorkspaceRootConfig,
 };
-pub use crate::util::toml::InheritableFields;
+pub use crate::util::toml::schema::InheritableFields;
 
 pub mod compiler;
 pub mod dependency;
