+++ tests/testsuite/profile_config.rs
@@ -1,6 +1,6 @@
 //! Tests for profiles defined in config files.
 
-use cargo::util::toml::TomlDebugInfo;
+use cargo::util::toml::schema::TomlDebugInfo;
 use cargo_test_support::paths::CargoPathExt;
 use cargo_test_support::registry::Package;
 use cargo_test_support::{basic_lib_manifest, paths, project};
